+++ lib/src/utils/log_message.dart
@@ -3,7 +3,7 @@ import 'dart:developer' as developer;
 import 'package:flutter/foundation.dart';
 
 void logMessage(String message, {StackTrace? stackTrace, Object? error}) {
-  if (kDebugMode) {
+  if (error != null || kDebugMode) {
     developer.log('[Video Player]: $message', stackTrace: stackTrace, error: error);
   }
 }